Organize code to simplify logic
- The ThreemaAdapter::Outbound is no longer an ActiveJob, but instead
  determines which job to enqueu based on the type of message we want to
send out.
