+++ spec/adapters/threema_adapter/outbound_spec.rb
@@ -3,7 +3,6 @@
 require 'rails_helper'
 
 RSpec.describe ThreemaAdapter::Outbound do
-  let(:adapter) { described_class.new }
   let(:threema_id) { 'V5EA564T' }
   let(:contributor) do
     build(:contributor, threema_id: threema_id, email: nil).tap { |contributor| contributor.save(validate: false) }
@@ -15,7 +14,7 @@ RSpec.describe ThreemaAdapter::Outbound do
 
     subject { -> { described_class.send!(message) } }
 
-    it { should enqueue_job(described_class) }
+    it { should enqueue_job(described_class::Text) }
 
     context 'contributor has no threema_id' do
       let(:contributor) { create(:contributor, threema_id: nil, email: nil) }
@@ -26,29 +25,11 @@ RSpec.describe ThreemaAdapter::Outbound do
   describe '::send_welcome_message!' do
     subject { -> { described_class.send_welcome_message!(contributor) } }
 
-    it { should enqueue_job(described_class) }
+    it { should enqueue_job(described_class::Text) }
 
     context 'contributor has no threema_id' do
       let(:contributor) { create(:contributor, threema_id: nil, email: nil) }
       it { should_not enqueue_job(described_class) }
-    end
-  end
-
-  describe '#perform' do
-    subject { -> { adapter.perform(text: message.text, recipient: message.recipient) } }
-
-    it 'sends the message' do
-      expect_any_instance_of(Threema).to receive(:send).with(type: :text, threema_id: 'V5EA564T', text: message.text)
-      subject.call
-    end
-
-    context 'with lowercase Threema ID' do
-      let(:threema_id) { 'v5ea564t' }
-
-      it 'converts ID to uppercase' do
-        expect_any_instance_of(Threema).to receive(:send).with(type: :text, threema_id: 'V5EA564T', text: message.text)
-        subject.call
-      end
     end
   end
